requirements.txt
pygame==2.5.2
Flask==3.0.0
supabase==2.3.0
python-dotenv==1.0.0
Pillow==10.2.0
pygame-gui==0.6.9
psutil==5.9.7
requests==2.31.0
watchdog==3.0.0
SQLAlchemy==2.0.25
Flask
Pillow
psutil
pygame
pygame-gui
python-dotenv
requests
SQLAlchemy
supabase
watchdog
